+++ frontend/tests/infrastructureData.test.ts
@@ -65,6 +65,9 @@ const map1 = [
     [1, 1, 1, 1, 1, 1, 1, 3],
 ]
 // Floor test data
+const footprintA = [0, 1, 2, 3];    // Left side
+const footprintB = [4, 5, 6, 7];    // Will link A and C
+const footprintC = [8, 9, 10, 11];  // Right side
 
 describe("Infrastructure Data", () => {
     const infraData = new InfrastructureData();
@@ -184,8 +187,10 @@ describe("Infrastructure Data", () => {
     })
 
     test("Can create a new floor", () => {
-        infraData.addNewFloor(5, [0, 1, 2, 3], infraData._currentSerial);
+        expect(infraData._currentSerial).toBe(1000);
+        infraData.addNewFloor(5, footprintA, infraData._currentSerial);
         expect(infraData._floors.length).toBe(1);
+        expect(infraData._currentSerial).toBe(1001);    // Validate serial number augmentation
     })
 
     test("Can find floors at a certain elevation", () => {
@@ -197,11 +202,36 @@ describe("Infrastructure Data", () => {
     
 
     test("Can delete a floor", () => {
-
+        // Delete all of the floors from the previous test (validates unique ID generator as well)
+        infraData._floors.forEach((floor) => {
+            infraData.deleteFloor(floor._id);
+        })
+        expect(infraData._floors.length).toBe(0);
     })
 
     test("Can combine two floors", ()  => {
-
+        // Reset serial number to target the right floor IDs
+        infraData.resetSerialNumber();
+        expect(infraData._currentSerial).toBe(1000);    // Validate serial number reset
+        // Ensure no previous floors exist
+        infraData._floors = [];
+        // Add two new floors that are on the same level, but with a gulf between them
+        infraData.addNewFloor(5, footprintA, 9001);
+        expect(infraData._floors[0]._id).toBe(1001);
+        infraData.addNewFloor(5, footprintC, 9002);
+        expect(infraData._floors.length).toBe(2);
+        expect(infraData._floors[0]._id).toBe(1001);
+        expect(infraData._floors[1]._id).toBe(1002);
+        // Validate pre-combine details for comparison
+        expect(infraData._floors[0]._modules).toStrictEqual([9001]);
+        expect(infraData._floors[0]._rightSide).toBe(3);
+        infraData.combineFloors(1001, 1002, 9003);
+        // Validate post-combine details
+        expect(infraData._floors.length).toBe(1);       // Only one floor exists
+        expect(infraData._floors[0]._id).toBe(1001);    // Its ID is that of the first floor created
+        expect(infraData._floors[0]._modules).toStrictEqual([9001, 9002, 9003]);  // Both of the existing floors' modules IDs are preserved, and the new ID of the new module that unites them is also present
+        expect(infraData._floors[0]._leftSide).toBe(0);
+        expect(infraData._floors[0]._rightSide).toBe(11);
     })